move some things out of executor, reach execution panic
