+++ riscv/tests/riscv_data/plonky3_recursion/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "plonky3_recursion"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+powdr-riscv-runtime = { path = "../../../../riscv-runtime", features = ["std", "getrandom"] }
+powdr-plonky3 = { path = "../../../../plonky3" }
+powdr-number = { path = "../../../../number" }
+indexmap = { version = "1.9.3", features = ["std"] } # enabling std on indexmap since somehow it thinks we do not have std
+
+[workspace]
