+++ riscv/tests/riscv_data/plonky3_recursion/src/main.rs
@@ -1,0 +1,6 @@
+extern crate powdr_riscv_runtime;
+use powdr_number::GoldilocksField;
+
+pub fn main() {
+    let _ = powdr_plonky3::verify::<GoldilocksField>(unimplemented!(), unimplemented!(), unimplemented!(), unimplemented!(), unimplemented!());
+}
